AI/plant-disease-ai/requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
numpy>=1.24.0
opencv-python>=4.8.0.76
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.2

# Serving
fastapi>=0.110.0
uvicorn>=0.23.1
python-multipart>=0.0.9
pydantic>=2.0

# Optional exports
onnx>=1.14.1
onnxruntime>=1.16.0

# (Optional) Detection if you plan to use YOLO later
# ultralytics>=8.0.0
